+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ### 🚩 Deprecations 🚩
 
+- Deprecate duplicate settings in service.ConfigProvider, embed ResolverSettings (#5843)
 - Deprecate `featuregate.Registry.MustApply` in favor of `featuregate.Registry.Apply`. (#5801)
 - Deprecate the `component.Factory.StabilityLevel(config.DataType)` in favor of Stability per component (#5762):
   - `component.ExporterFactory.TracesExporterStability`